Program.cs adn Test Methods are updated.
