+++ csharp-advanced-topics/FuncAndAction/FuncAndAction/Program.cs
@@ -12,17 +12,20 @@ action(123);
 
 // Also we can use this calling format
 action.Invoke(123);
-partial class Program
+
+public partial class Program
 {
     public delegate int DelegateMethod(int a, int b);
     public static int DisplayResult(int a, int b)
     {
         return a + b;
     }
+
     public static int CalculateValue(int a, int b)
     {
         return a * b;
     }
+
     public static void PrintValue(int number)
     {
         Console.WriteLine(number);
