+++ csharp-advanced-topics/FuncAndAction/FuncAndActionTest/FuncAndActionTest.csproj
@@ -20,4 +20,8 @@
     </PackageReference>
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\FuncAndAction\FuncAndAction.csproj" />
+  </ItemGroup>
+
 </Project>
